# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for anonymous random voice/video matching and WebSocket signaling.",
        "title": "Random Match API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/auth/anonymous": {
            "post": {
                "description": "Creates an anonymous user and returns a JWT for API access.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Create anonymous user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.anonymousAuthResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/session": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Verifies the bearer token and returns the authenticated user id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify current session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.authSessionResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/match/join": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds the current user to a voice or video matchmaking queue. Returns 202 while waiting or 200 when matched.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match"
                ],
                "summary": "Join matchmaking queue",
                "parameters": [
                    {
                        "description": "Match preferences",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_server.joinMatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.matchedResponse"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/internal_server.waitingMatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/match/leave": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marks the current user as having left matchmaking.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match"
                ],
                "summary": "Leave matchmaking queue",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.leaveMatchResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/match/snapshot": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Saves a camera snapshot on the server after a match is created.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match"
                ],
                "summary": "Save match snapshot",
                "parameters": [
                    {
                        "description": "Snapshot payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_server.snapshotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.snapshotResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/stats": {
            "get": {
                "description": "Returns current online users, waiting users, and users in active chats.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Runtime usage stats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.statsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/ws": {
            "get": {
                "description": "Upgrades to a WebSocket used for match and WebRTC signaling messages. Pass the JWT as the token query parameter.",
                "tags": [
                    "signaling"
                ],
                "summary": "Connect signaling WebSocket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns the backend service health status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.healthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_server.anonymousAuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "$ref": "#/definitions/random-match_backend_internal_model.User"
                }
            }
        },
        "internal_server.authSessionResponse": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "string",
                    "example": "4f6d8c2a7b1e4a9f8c0d2e3f5a6b7c8d"
                }
            }
        },
        "internal_server.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "unauthorized"
                }
            }
        },
        "internal_server.healthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "internal_server.joinMatchRequest": {
            "type": "object",
            "properties": {
                "mode": {
                    "enum": [
                        "video",
                        "voice"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/random-match_backend_internal_model.MatchMode"
                        }
                    ],
                    "example": "video"
                },
                "region": {
                    "type": "string",
                    "example": "global"
                }
            }
        },
        "internal_server.leaveMatchResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "left"
                }
            }
        },
        "internal_server.matchedResponse": {
            "type": "object",
            "properties": {
                "initiator": {
                    "type": "boolean",
                    "example": true
                },
                "peerId": {
                    "type": "string",
                    "example": "5f6d8c2a7b1e4a9f8c0d2e3f5a6b7c8d"
                },
                "roomId": {
                    "type": "string",
                    "example": "f9b9fdc7110a4a5cb6924f2d936cd58a"
                },
                "status": {
                    "type": "string",
                    "example": "matched"
                }
            }
        },
        "internal_server.snapshotRequest": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "integer",
                    "example": 480
                },
                "image": {
                    "type": "string",
                    "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQ..."
                },
                "mode": {
                    "type": "string",
                    "example": "video"
                },
                "peerId": {
                    "type": "string",
                    "example": "5f6d8c2a7b1e4a9f8c0d2e3f5a6b7c8d"
                },
                "roomId": {
                    "type": "string",
                    "example": "f9b9fdc7110a4a5cb6924f2d936cd58a"
                },
                "width": {
                    "type": "integer",
                    "example": 640
                }
            }
        },
        "internal_server.snapshotResponse": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string",
                    "example": "/app/snapshots/2026-05-22/room-user.jpg"
                },
                "status": {
                    "type": "string",
                    "example": "saved"
                }
            }
        },
        "internal_server.statsResponse": {
            "type": "object",
            "properties": {
                "chatting": {
                    "type": "integer",
                    "example": 8
                },
                "online": {
                    "type": "integer",
                    "example": 12
                },
                "waiting": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "internal_server.waitingMatchResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "waiting"
                }
            }
        },
        "random-match_backend_internal_model.MatchMode": {
            "type": "string",
            "enum": [
                "video",
                "voice"
            ],
            "x-enum-varnames": [
                "MatchModeVideo",
                "MatchModeVoice"
            ]
        },
        "random-match_backend_internal_model.User": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "firebaseUid": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by a JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}